backend/requirements.txt
fastapi
uvicorn
pydantic
db-sqlite3
google-generativeai
websockets
python-dotenv